requirements.txt
chainlit==1.0.0
langchain==0.2.5
langchain_community==0.2.5
langchain_core==0.2.9
langchain_huggingface==0.0.3
langchain_text_splitters==0.2.1
langchain_openai==0.1.9
python-dotenv==1.0.1
pymupdf==1.24.5
faiss-cpu==1.8.0.post1
qdrant-client==1.9.2